+++ be/src/vec/data_types/serde/data_type_decimal_serde.cpp
@@ -85,8 +85,9 @@ Status DataTypeDecimalSerDe<T>::deserialize_one_cell_from_json(IColumn& column, 
 
 template <typename T>
 void DataTypeDecimalSerDe<T>::write_column_to_arrow(const IColumn& column, const NullMap* null_map,
-                                                    arrow::ArrayBuilder* array_builder, int start,
-                                                    int end, const cctz::time_zone& ctz) const {
+                                                    arrow::ArrayBuilder* array_builder,
+                                                    int64_t start, int64_t end,
+                                                    const cctz::time_zone& ctz) const {
     auto& col = reinterpret_cast<const ColumnDecimal<T>&>(column);
     if constexpr (std::is_same_v<T, Decimal<Int128>>) {
         auto& builder = reinterpret_cast<arrow::Decimal128Builder&>(*array_builder);
