+++ be/src/vec/data_types/serde/data_type_ipv4_serde.cpp
@@ -123,8 +123,8 @@ Status DataTypeIPv4SerDe::read_column_from_pb(IColumn& column, const PValues& ar
 }
 
 void DataTypeIPv4SerDe::write_column_to_arrow(const IColumn& column, const NullMap* null_map,
-                                              arrow::ArrayBuilder* array_builder, int start,
-                                              int end, const cctz::time_zone& ctz) const {
+                                              arrow::ArrayBuilder* array_builder, int64_t start,
+                                              int64_t end, const cctz::time_zone& ctz) const {
     const auto& col_data = assert_cast<const ColumnIPv4&>(column).get_data();
     auto& int32_builder = assert_cast<arrow::Int32Builder&>(*array_builder);
     auto arrow_null_map = revert_null_map(null_map, start, end);
