+++ be/src/vec/data_types/serde/data_type_ipv4_serde.h
@@ -56,7 +56,7 @@ public:
                               int end) const override;
     Status read_column_from_pb(IColumn& column, const PValues& arg) const override;
     void write_column_to_arrow(const IColumn& column, const NullMap* null_map,
-                               arrow::ArrayBuilder* array_builder, int start, int end,
+                               arrow::ArrayBuilder* array_builder, int64_t start, int64_t end,
                                const cctz::time_zone& ctz) const override;
     void read_column_from_arrow(IColumn& column, const arrow::Array* arrow_array, int64_t start,
                                 int64_t end, const cctz::time_zone& ctz) const override;
